improved canceler context
