+++ framework/framework.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/micrologger"
 	"k8s.io/client-go/tools/cache"
 
-	"github.com/giantswarm/operatorkit/framework/cancelercontext"
+	"github.com/giantswarm/operatorkit/framework/canceledcontext"
 )
 
 // Config represents the configuration used to create a new operator framework.
@@ -73,7 +73,7 @@ func (f *Framework) AddFunc(obj interface{}) {
 	defer f.mutex.Unlock()
 
 	ctx := context.Background()
-	ctx = cancelercontext.NewContext(ctx, make(chan struct{}))
+	ctx = canceledcontext.NewContext(ctx, make(chan struct{}))
 
 	f.logger.Log("action", "start", "component", "operatorkit", "function", "ProcessCreate")
 
@@ -107,7 +107,7 @@ func (f *Framework) DeleteFunc(obj interface{}) {
 	defer f.mutex.Unlock()
 
 	ctx := context.Background()
-	ctx = cancelercontext.NewContext(ctx, make(chan struct{}))
+	ctx = canceledcontext.NewContext(ctx, make(chan struct{}))
 
 	f.logger.Log("action", "start", "component", "operatorkit", "function", "ProcessDelete")
 
@@ -161,7 +161,7 @@ func ProcessCreate(ctx context.Context, obj interface{}, resources []Resource) e
 	}
 
 	for _, r := range resources {
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		currentState, err := r.GetCurrentState(ctx, obj)
@@ -169,7 +169,7 @@ func ProcessCreate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		desiredState, err := r.GetDesiredState(ctx, obj)
@@ -177,7 +177,7 @@ func ProcessCreate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		createState, err := r.GetCreateState(ctx, obj, currentState, desiredState)
@@ -185,7 +185,7 @@ func ProcessCreate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		err = r.ProcessCreateState(ctx, obj, createState)
@@ -219,7 +219,7 @@ func ProcessDelete(ctx context.Context, obj interface{}, resources []Resource) e
 	}
 
 	for _, r := range resources {
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		currentState, err := r.GetCurrentState(ctx, obj)
@@ -227,7 +227,7 @@ func ProcessDelete(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		desiredState, err := r.GetDesiredState(ctx, obj)
@@ -235,7 +235,7 @@ func ProcessDelete(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		deleteState, err := r.GetDeleteState(ctx, obj, currentState, desiredState)
@@ -243,7 +243,7 @@ func ProcessDelete(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		err = r.ProcessDeleteState(ctx, obj, deleteState)
@@ -278,7 +278,7 @@ func ProcessUpdate(ctx context.Context, obj interface{}, resources []Resource) e
 	}
 
 	for _, r := range resources {
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		currentState, err := r.GetCurrentState(ctx, obj)
@@ -286,7 +286,7 @@ func ProcessUpdate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		desiredState, err := r.GetDesiredState(ctx, obj)
@@ -294,7 +294,7 @@ func ProcessUpdate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		createState, deleteState, updateState, err := r.GetUpdateState(ctx, obj, currentState, desiredState)
@@ -302,7 +302,7 @@ func ProcessUpdate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		err = r.ProcessCreateState(ctx, obj, createState)
@@ -310,7 +310,7 @@ func ProcessUpdate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		err = r.ProcessDeleteState(ctx, obj, deleteState)
@@ -318,7 +318,7 @@ func ProcessUpdate(ctx context.Context, obj interface{}, resources []Resource) e
 			return microerror.Mask(err)
 		}
 
-		if cancelercontext.IsCanceled(ctx) {
+		if canceledcontext.IsCanceled(ctx) {
 			return nil
 		}
 		err = r.ProcessUpdateState(ctx, obj, updateState)
